Add next-dev and next-start test mode handling

+++ test/lib/next-modes/base.ts
@@ -1,11 +1,13 @@
 import os from 'os'
 import path from 'path'
 import fs from 'fs-extra'
+import treeKill from 'tree-kill'
 import { NextConfig } from 'next'
 import { FileRef } from '../e2e-utils'
+import { ChildProcess } from 'child_process'
 import { createNextInstall } from '../create-next-install'
 
-type Event = 'stdout' | 'stderr' | 'error'
+type Event = 'stdout' | 'stderr' | 'error' | 'destroy'
 
 export class NextInstance {
   protected files: {
@@ -15,6 +17,8 @@ export class NextInstance {
   protected dependencies?: { [name: string]: string }
   protected events: { [eventName: string]: Set<any> }
   protected testDir: string
+  protected isDestroyed: boolean
+  protected childProcess: ChildProcess
 
   constructor({
     files,
@@ -33,18 +37,25 @@ export class NextInstance {
     this.dependencies = dependencies
     this.nextConfig = nextConfig
     this.events = {}
+    this.isDestroyed = false
   }
 
   protected async createTestDir() {
+    if (this.isDestroyed) {
+      throw new Error('next instance already destroyed')
+    }
+    console.log(`Creating test directory with isolated next...`)
+
     const tmpDir = process.env.NEXT_TEST_DIR || (await fs.realpath(os.tmpdir()))
     this.testDir = path.join(tmpDir, `next-test-${Date.now()}`)
 
-    if (process.env.NEXT_TEST_STARTER) {
+    if (process.env.NEXT_TEST_STARTER && !this.dependencies) {
       await fs.copy(process.env.NEXT_TEST_STARTER, this.testDir)
     } else {
       this.testDir = await createNextInstall({
         react: 'latest',
         'react-dom': 'latest',
+        ...this.dependencies,
       })
     }
 
@@ -53,17 +64,48 @@ export class NextInstance {
       const outputFilename = path.join(this.testDir, filename)
 
       if (typeof item === 'string') {
+        await fs.ensureDir(path.dirname(outputFilename))
         await fs.writeFile(outputFilename, item)
       } else {
         await fs.copy(item.fsPath, outputFilename)
       }
     }
+    console.log(`Test directory created at ${this.testDir}`)
   }
 
   public async setup(): Promise<void> {}
   public async start(): Promise<void> {}
+  public async stop(): Promise<void> {
+    if (this.childProcess) {
+      let exitResolve
+      const exitPromise = new Promise((resolve) => {
+        exitResolve = resolve
+      })
+      this.childProcess.addListener('exit', () => {
+        exitResolve()
+      })
+      await new Promise<void>((resolve) => {
+        treeKill(this.childProcess.pid, 'SIGKILL', (err) => {
+          if (err) {
+            console.error('tree-kill', err)
+          }
+          resolve()
+        })
+      })
+      this.childProcess.kill('SIGKILL')
+      await exitPromise
+      this.childProcess = undefined
+      console.log(`Stopped next server`)
+    }
+  }
 
   public async destroy(): Promise<void> {
+    if (this.isDestroyed) {
+      throw new Error(`next instance already destroyed`)
+    }
+    this.isDestroyed = true
+    await this.stop()
+
     if (this.dependencies || !process.env.NEXT_TEST_STARTER) {
       await fs.remove(this.testDir)
     } else {
@@ -75,6 +117,7 @@ export class NextInstance {
         }
       }
     }
+    this.emit('destroy', [])
   }
 
   public url(): string {
